requirements.txt
aiokafka==0.10.0
asyncpg==0.29.0
fastapi==0.111.0
numpy==1.22.0
opencv_python==4.9.0.80
PyYAML==6.0
PyYAML==6.0.1
PyYAML==6.0.1
SQLAlchemy==2.0.30
ulid_py==1.1.0
uvicorn==0.30.0
ultralytics